UHF-11682: Move paatokset_submenus code into paatokset_ahjo_api module
