+++ conf/cmi/core.extension.yml
@@ -137,7 +137,6 @@ module:
   paatokset_policymakers: 0
   paatokset_search: 0
   paatokset_search_form: 0
-  paatokset_submenus: 0
   page_cache: 0
   paragraphs_asymmetric_translation_widgets: 0
   paragraphs_library: 0
