+++ public/modules/custom/paatokset_ahjo_api/src/Plugin/Block/DocumentsBlock.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Drupal\paatokset_submenus\Plugin\Block;
+namespace Drupal\paatokset_ahjo_api\Plugin\Block;
 
 use Drupal\Core\Block\Attribute\Block;
 use Drupal\Core\Block\BlockBase;
@@ -13,42 +13,52 @@ use Drupal\paatokset_policymakers\Service\PolicymakerService;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 
 /**
- * Provides Agendas Submenu Block.
+ * Provides Agendas Submenu Documents Block.
  */
 #[Block(
-  id: 'agendas_submenu',
-  admin_label: new TranslatableMarkup('Agendas Submenu'),
+  id: 'agendas_submenu_documents',
+  admin_label: new TranslatableMarkup('Paatokset policymaker documents'),
   category: new TranslatableMarkup('Paatokset custom blocks'),
 )]
-final class AgendasSubmenuBlock extends BlockBase implements ContainerFactoryPluginInterface {
+class DocumentsBlock extends BlockBase implements ContainerFactoryPluginInterface {
 
   /**
-   * The policymaker service.
+   * {@inheritDoc}
    */
-  private PolicymakerService $policymakerService;
+  final public function __construct(
+    array $configuration,
+    $plugin_id,
+    $plugin_definition,
+    private readonly PolicymakerService $policymakerService,
+  ) {
+    parent::__construct($configuration, $plugin_id, $plugin_definition);
+    $this->policymakerService->setPolicyMakerByPath();
+  }
 
   /**
    * {@inheritdoc}
    */
   public static function create(ContainerInterface $container, array $configuration, $plugin_id, $plugin_definition): static {
-    $plugin = new self($configuration, $plugin_id, $plugin_definition);
-    $plugin->policymakerService = $container->get('paatokset_policymakers');
-    return $plugin;
+    return new static(
+      $configuration,
+      $plugin_id,
+      $plugin_definition,
+      $container->get('paatokset_policymakers')
+    );
   }
 
   /**
    * {@inheritDoc}
    */
   public function build(): array {
-    $list = $this->policymakerService->getAgendasListFromElasticSearch(NULL, TRUE);
-    $years = array_keys($list);
+    $list = $this->policymakerService->getApiMinutesFromElasticSearch(NULL, TRUE);
 
     return [
       '#theme' => 'agendas_submenu',
-      '#title' => 'Viranhaltijapäätökset',
-      '#years' => $years,
+      '#title' => $this->t('Office holder decisions'),
+      '#years' => array_keys($list),
       '#list' => $list,
-      '#type' => 'decisions',
+      '#type' => 'documents',
     ];
   }
 
@@ -59,7 +69,7 @@ final class AgendasSubmenuBlock extends BlockBase implements ContainerFactoryPlu
     $policymaker = $this->policymakerService->getPolicyMaker();
     if ($policymaker instanceof NodeInterface && $policymaker->hasField('field_policymaker_id')) {
       $policymaker_id = $policymaker->get('field_policymaker_id')->value;
-      return ['decision_pm:' . $policymaker_id];
+      return ['meeting_pm:' . $policymaker_id];
     }
     return [];
   }
