--- public/modules/custom/paatokset_submenus/config/install/system.menu.policymaker-navigation.yml
@@ -1,7 +1,0 @@
-langcode: und
-status: true
-dependencies: {  }
-id: policymaker-navigation
-label: 'Navigation for policymaker pages'
-description: 'Navigation for policymaker pages'
-locked: false
