+++ public/modules/custom/paatokset_submenus/paatokset_submenus.info.yml
@@ -1,4 +1,4 @@
-name: Paatokset meetings submenu
+name: Paatokset meetings submenu (DEPRECATED)
 type: module
 package: custom
 core_version_requirement: ^9 || ^10
