--- public/modules/custom/paatokset_submenus/paatokset_submenus.module
@@ -1,36 +1,0 @@
-<?php
-
-/**
- * @file
- * Contains paatokset_submenus.
- */
-
-declare(strict_types=1);
-
-/**
- * Implements hook_theme().
- */
-function paatokset_submenus_theme($existing, $type, $theme, $path): array {
-  return [
-    'policymaker_side_navigation' => [
-      'variables' => [
-        'items' => [],
-        'menu_attributes' => NULL,
-        'menu_link_parent' => NULL,
-      ],
-    ],
-    'policymaker_side_navigation_mobile' => [
-      'variables' => [
-        'items' => [],
-        'current_option' => NULL,
-      ],
-    ],
-    'agendas_submenu' => [
-      'variables' => [
-        'years' => [],
-        'list' => [],
-        'type' => NULL,
-      ],
-    ],
-  ];
-}
